refactor(extension-examples): rename type.ts to index.ts
