+++ examples/typescript-basic/frontity.settings.ts
@@ -1,7 +1,7 @@
 import { Settings } from "frontity/types";
-import ExtensionExample1 from "@frontity/extension-example-1/type";
-import ExtensionExample2 from "@frontity/extension-example-2/type";
-import TinyRouter from "@frontity/tiny-router/type";
+import ExtensionExample1 from "@frontity/extension-example-1";
+import ExtensionExample2 from "@frontity/extension-example-2";
+import TinyRouter from "@frontity/tiny-router";
 
 const settings: Settings<ExtensionExample1 | ExtensionExample2 | TinyRouter> = [
   {
