+++ packages/extension-example-2/index.ts
